Rework pseudoVersionToRpmVersion to parse more versions
A few of the dependencies of odo could not be parsed by
pseudoVersionToRpmVersion
This commit reworks this parser in order to support them.
Instead of splitting the version string on '-', and matching
each substring separately, we know always work on the full string,
removing the parts which were successfully matched.
This is done as the 'prerelease' part of the version seems to be
possibly very arbitrary, and can contain '-'. So we now match version at
the beginning, commit/date at the end of the string, and assume the rest
to be a prerelease string.